reorganized icons
