+++ src/icons/CollapsibleListIcons/PrivateChannelIcon.tsx
