+++ src/icons/CollapsibleListIcons/ToggleShowChannelsIcon.tsx
